# backend/openapi.json
{
  "basePath": "/",
...
  "info": {
    "description": "API documentation for IRIS backend.",
    "title": "IRIS - Intelligent Resume Insight System API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/jobs": {
      "get": {
        "parameters": [
          {
            "example": "Python",
            "in": "query",
            "name": "title",
            "required": false,
            "type": "string"
          },
          {
            "example": "Bangalore",
            "in": "query",
            "name": "location",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Job list",
            "examples": {
              "application/json": [
                {
                  "created_at": "2026-02-22T10:00:00",
                  "description": "Build Flask APIs",
                  "employer_id": 2,
                  "id": 1,
                  "location": "Bangalore",
                  "required_skills": [
                    "flask",
                    "python"
                  ],
                  "title": "Python Backend Developer"
                }
              ]
            }
          }
        },
        "summary": "List jobs with optional search filters",
        "tags": [
          "Jobs"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "description": {
                  "example": "Build Flask APIs and optimize SQL queries",
                  "type": "string"
                },
                "location": {
                  "example": "Bangalore",
                  "type": "string"
                },
                "required_skills": {
                  "example": [
                    "python",
                    "flask",
                    "mysql",
                    "sql"
                  ],
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "title": {
                  "example": "Python Backend Developer",
                  "type": "string"
                }
              },
              "required": [
                "title",
                "description",
                "location",
                "required_skills"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Job created",
            "examples": {
              "application/json": {
                "job_id": 1,
                "message": "Job created successfully"
              }
            }
          },
          "403": {
            "description": "Only employer can create jobs"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create job (Employer)",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/api/jobs/{job_id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "examples": {
              "application/json": {
                "message": "Job deleted successfully"
              }
            }
          },
          "404": {
            "description": "Job not found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete job (Employer)",
        "tags": [
          "Jobs"
        ]
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "description": {
                  "type": "string"
                },
                "location": {
                  "type": "string"
                },
                "required_skills": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "title": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "examples": {
              "application/json": {
                "message": "Job updated successfully"
              }
            }
          },
          "404": {
            "description": "Job not found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Update job (Employer)",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/api/jobs/{job_id}/applicants": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Applicants list",
            "examples": {
              "application/json": {
                "applicants": [
                  {
                    "application_id": 5,
                    "applied_at": "2026-02-22T12:00:00",
                    "email": "john@x.com",
                    "match_score": 88.4,
                    "name": "John",
                    "status": "Applied",
                    "user_id": 7
                  }
                ],
                "job_id": 1
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "View applicants by job sorted by match score",
        "tags": [
          "Jobs"
        ]
      }
    },
...
  },
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "description": "JWT format: Bearer <token>",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}